fix: deleteRange start must be > 0 in _deleteBackward (#928)
* fix: _deleteBackward deleteRange start must > 0

* fix: _deleteBackward deleteRange start must > 0

+++ lib/src/editor/editor_component/service/ime/non_delta_input_service.dart
@@ -215,6 +215,11 @@ class NonDeltaTextInputService extends TextInputService with TextInputClient {
         baseOffset: selection.baseOffset - 1,
         extentOffset: selection.extentOffset - 1,
       );
+
+      if (!deleteRange.isValid) {
+        return;
+      }
+
       // valid the result
       onDelete(
         TextEditingDeltaDeletion(
